Сделал заглушки репозиториев, main, подключил схему SQL

+++ src/main/java/org/course/aston/entity/Fractions.java
@@ -1,15 +1,16 @@
 package org.course.aston.entity;
 
 public class Fractions {
-    private int fractionID;
+
+    private Long fractionID;
     private String fractionName;
 
-    public Fractions(int fractionID, String fractionName) {
+    public Fractions(Long fractionID, String fractionName) {
         this.fractionID = fractionID;
         this.fractionName = fractionName;
     }
 
-    public int getFractionID() {
+    public Long getFractionID() {
         return fractionID;
     }
