+++ src/main/java/org/course/aston/entity/Heroes.java
@@ -1,12 +1,12 @@
 package org.course.aston.entity;
 
 public class Heroes {
-    private int heroId;
+    private Long heroId;
     private String heroName;
     private String heroLastname;
     private int roleNameID;
 
-    public Heroes(int heroId, String heroName, String heroLastname, int roleNameID) {
+    public Heroes(Long heroId, String heroName, String heroLastname, int roleNameID) {
         this.heroId = heroId;
         this.heroName = heroName;
         this.heroLastname = heroLastname;
